♻️ refactor: 친구 추가 로직 분리리
